Move base64 to separate file.
